+++ src/components/translate/TranslationPanel.vue
@@ -2,8 +2,12 @@
 
 <script setup lang="ts">
 import Communication from './Communication.vue';
+import { NInput, NInputGroup, NButton, NIcon, NTooltip } from 'naive-ui';
+import {  LogoGithub, Raw, VolumeFileStorage } from '@vicons/carbon'
+import { HistoryFilled } from '@vicons/material'
 import { ref, onMounted, nextTick, computed, watch } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
+import { DataSourceType } from '../../store'
 
 
 const route = useRoute()
@@ -15,13 +19,16 @@ const query = computed(()=>{
 
 let csvUrl = ref('');
 let communication = ref<InstanceType<typeof Communication> | null>(null);
+let urlInput = ref<InstanceType<typeof NInput> | null>(null);
+const currentMode = ref("" as DataSourceType)
 
 // if this page is never loaded, onMounted will activate
 onMounted(()=>{
   removeForceReloadParamInSearch()
   const mode = (new URLSearchParams(window.location.search)).get("mode")
-  if (mode === "storage") {
+  if (mode === "history") {
     if(location.hash) {
+      currentMode.value = DataSourceType.History
       location.hash = decodeURIComponent(location.hash)
       const name = decodeURIComponent(location.hash.substring(1))
       nextTick(()=>communication.value?.loadDataFromLocalStorage(name))
@@ -30,7 +37,7 @@ onMounted(()=>{
     if(location.hash) {
       location.hash = decodeURIComponent(location.hash)
       csvUrl.value = decodeURIComponent(location.hash.substring(1))
-      nextTick(()=>communication.value?.loadDataFromUrl(csvUrl.value))
+      nextTick(()=>loadDataFromUrl(csvUrl.value))
     }
   }
   
@@ -57,10 +64,28 @@ function confirm() {
   loadDataFromUrl(csvUrl.value)
 }
 
-function loadDataFromUrl(url: string) {
-  location.hash = url
-  location.search = ""
-  communication.value?.loadDataFromUrl(url);
+async function loadDataFromUrl(url: string) {
+  if (url.endsWith(".csv")) {
+    currentMode.value = DataSourceType.Server
+    await communication.value?.loadDataFromGithubCsvUrl(url)
+  } else if (url.endsWith(".json")) {
+    currentMode.value = DataSourceType.Raw
+    await communication.value?.loadDataFromJsonPathUrl(url)
+  } else {
+    console.log(url)
+    alert("unexpected url: should ends with .csv or .json")
+  }
+  router.replace({
+    path: route.path,
+    hash: `#${url}`,
+    query: {
+      mode: currentMode.value.toString()
+    }
+  })
+  nextTick(()=> urlInput.value?.scrollTo({
+    behavior: 'smooth',
+    left: 10000
+  }))
 }
 
 function handleFileChange(e: Event) {
@@ -71,54 +96,72 @@ function handleFileChange(e: Event) {
     if (communication.value !== null) {
       csvUrl.value = ""
       window.location.hash = ""
+      currentMode.value = DataSourceType.File
       communication.value?.loadDataFromFile(file)
     }
   }
 }
 
-function toGithub() {
-  if (csvUrl.value) {
-    window.open(csvUrl.value)
-  }
-  else {
-    window.open("https://github.com/ShinyGroup/SCTranslationData/tree/master/data/story")
-  }
-}
+// function toGithub() {
+//   if (csvUrl.value) {
+//     window.open(csvUrl.value)
+//   }
+//   else {
+//     window.open("https://github.com/ShinyGroup/SCTranslationData/tree/master/data/story")
+//   }
+// }
 
-let iconSrc = computed(()=>{
-  return "github.png"
-})
+// let iconSrc = computed(()=>{
+//   return "github.png"
+// })
 
 </script>
 <template>
     <div class="input-row">
-      <input v-model="csvUrl" placeholder="Enter json path or Github CSV URL" class="url-input" @keypress.enter="confirm"/>
-      <button @click="confirm">Confirm</button>
-      <img class="github" :src="iconSrc" @click="toGithub" />
+      <n-input-group class="url-input">
+        <n-input v-model:value="csvUrl" placeholder="Enter json path or Github CSV URL"  @keypress.enter="confirm" clearable ref="urlInput" />
+        <n-button type="info" @click="confirm">Confirm</n-button>
+        <n-tooltip :show-arrow="false" trigger="hover" v-if="!!currentMode">
+          <template #trigger>
+            <n-button tertiary type="default">
+              <n-icon size="30">
+                <HistoryFilled v-if="currentMode==='history'" />
+                <LogoGithub v-if="currentMode==='server'"/>
+                <Raw v-if="currentMode==='raw'"/>
+                <VolumeFileStorage v-if="currentMode==='file'"/>
+              </n-icon>
+            </n-button>
+          </template>
+          Current mode: {{ currentMode }}
+        </n-tooltip>
+        
+      </n-input-group>
     </div>
     <div class="input-row">
       <label> or </label>
       <input type="file" @change="handleFileChange" class="file-selector"/>
     </div>
+    <!-- event from chapter change -->
     <Communication :csvUrl="csvUrl" ref="communication" @load-data="loadDataFromUrl"/>
 </template>
 <style scoped>
 .input-row {
   display: flex;
-  flex: 8;
+  /* flex: 8; */
   align-items: center;
+  text-align: left;
 }
 
-.input-row > .url-input {
+/* .input-row > .url-input {
   flex: 5;
   height: 100%;
   font-size: medium;
-}
+} */
 
-.github {
+/* .github {
   height: 30px;
   cursor: pointer;
-}
+} */
 
 /* .file-selector {
   appearance: none;